(WIP) async import use case
This is a patch containing all the code necessary to do an import of
an image. Not intended to go in as it is and would be split into
multiple patches shortly after FF ends and ported on top of master.
It is also helping to maintain the code without breaking the depending
patches when a change to later one is proposed.

wip for bp async-glance-workers

Change-Id: I879eebcb8d3e0c85db02463b0e6c28d43a90c666